refactor: 구입 금액 검증 및 금액 변환 부분 domain -> service 폴더로 이동
