+++ src/test/java/lotto/AmoutValidatorTest.java
@@ -1,0 +1,4 @@
+package lotto;
+
+public class AmoutValidatorTest {
+}
